requirements.txt
SpeechRecognition
pydub
logging
pyTelegramBotAPI
python-dotenv
